+++ pp/name.py
@@ -165,12 +165,13 @@ def clean_value(value: Any) -> str:
 
 def get_name(component_type: str, name: str) -> str:
     """Returns name with correct number of characters"""
-    if not isinstance(component_type, str):
-        raise ValueError(f"{component_type} needs to be a sting")
     if not isinstance(name, str):
-        raise ValueError(f"{name} needs to be a sting")
+        raise ValueError(f"{name} needs to be a string")
+    if not isinstance(component_type, str):
+        raise ValueError(f"{component_type} needs to be a string")
     if len(name) > MAX_NAME_LENGTH:
-        name = f"{component_type[:(MAX_NAME_LENGTH - 8)]}_{hashlib.md5(name.encode()).hexdigest()[:8]}"
+        name_hash = hashlib.md5(name.encode()).hexdigest()[:8]
+        name = f"{component_type[:(MAX_NAME_LENGTH - 8)]}_{name_hash}"
     return clean_name(name)